Remove inmemory exporter artifact, move to sdk-tracing
