--- exporters/inmemory/README.md
@@ -1,8 +1,0 @@
-# OpenTelemetry - In-Memory Exporter
-
-[![Javadocs][javadoc-image]][javadoc-url]
-
-* Java 8 compatible.
-
-[javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-exporters-inmemory.svg
-[javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-exporters-inmemory
