--- exporters/inmemory/build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id "java"
-    id "maven-publish"
-
-    id "ru.vyarus.animalsniffer"
-}
-
-description = 'OpenTelemetry InMemory Export'
-ext.moduleName = "io.opentelemetry.exporter.inmemory"
-
-dependencies {
-    api project(':opentelemetry-sdk')
-
-    annotationProcessor libraries.auto_value
-
-    testImplementation project(':opentelemetry-sdk-testing')
-
-    signature "org.codehaus.mojo.signature:java18:1.0@signature"
-    signature "net.sf.androidscents.signature:android-api-level-24:7.0_r2@signature"
-}
