+++ opentracing-shim/build.gradle
@@ -11,7 +11,6 @@ dependencies {
             libraries.opentracing
 
     testImplementation project(':opentelemetry-sdk-testing'),
-            project(':opentelemetry-exporter-inmemory'),
             libraries.junit,
             libraries.assertj,
             libraries.slf4jsimple
