+++ sdk/testing/build.gradle
@@ -12,6 +12,4 @@ dependencies {
     api libraries.junit_jupiter_api
 
     annotationProcessor libraries.auto_value
-
-    implementation project(':opentelemetry-exporter-inmemory')
 }
