+++ sdk/testing/src/test/java/io/opentelemetry/sdk/testing/exporter/InMemoryMetricExporterTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.exporter.inmemory;
+package io.opentelemetry.sdk.testing.exporter;
 
 import static org.assertj.core.api.Assertions.assertThat;
