+++ settings.gradle
@@ -28,7 +28,6 @@ include ":opentelemetry-all",
         ":opentelemetry-extension-runtime-metrics",
         ":opentelemetry-extension-trace-propagators",
         ":opentelemetry-extension-trace-utils",
-        ":opentelemetry-exporter-inmemory",
         ":opentelemetry-exporter-jaeger",
         ":opentelemetry-exporter-jaeger-thrift",
         ":opentelemetry-exporter-logging",
